requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
azure-cognitiveservices-speech==1.40.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
fastapi==0.112.4
ffmpeg==1.4
ffmpy==0.4.0
ffprobe==0.5
fsspec==2024.9.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.1
openai==1.47.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
ruff==0.6.7
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
decorator==5.1.1
docstring_parser==0.16
executing==2.1.0
filelock==3.16.1
fonttools==4.54.0
frozenlist==1.4.1
gradio==4.44.0
instructor==1.4.3
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
multidict==6.1.0
parso==0.8.4
prompt_toolkit==3.0.47
pure_eval==0.2.3
stack-data==0.6.3
tenacity==8.5.0
traitlets==5.14.3
wcwidth==0.2.13
yarl==1.12.1
